Use a `Config` file to store the `BEARER_TOKEN` values
It's been suggested that we don't tie our configuration for the bearer
tokens to the system (read: infrastructure). Instead we want to be
able to deploy the app and the localised configuration easily.

The decision is to use a config file as that's the simplest thing for
our current deployment setup.