Run performance platform HTTP requests in parallel
This change makes HTTP requests against the performance platform in
parallel instead of in serial.

For each request we're making we create a new goroutine for it and add
it to a `sync.WaitGroup`. The `WaitGroup` is used to block until all
goroutines have completed their work before continuing.

This seems to be the sanest way to complete work like this in parallel
where a permanent goroutine isn't required and the functions
themselves don't run based on unbounded work coming in.