Create aggregate statistics about info pages.
Gather PP data on all pages with problem reports or search
queries. Aggregate and normalise this data by the number of
unique page views for each page. Push the aggregated data
back to the PP.

This should eventually be run as a regular job, perhaps once
a week. We will then be able to generate regular reports
for the content team, helping them to fix problematic pages.

Has some messy code to deal with smart answers. This is
because GA aggregate page view and search term statistics
across all pages in a smart answer, but Feedex reports
problem reports separately for each page in a smart answer.
It would be good to deal with this when the data is
pushed into the PP, rather than here - that would also
simplify info-frontend.

We need to discuss how often, and where, to run the regular
job to gather these statistics. I haven't added anything
to do this yet. Nor have I added setup for this code into
the main Makefile.