Add timeouts to calls to upstream services
Use the http.DefaultClient to talk to things, rather than more than one
client.

This means that the connections to the Need API, Content API and
Performance Platform will all be using the same http.Client instance.
Previously, the Performance Platform was using the DefaultClient, and a
separate instance was being used to talk to the other two.

Change the timeout to be 2 seconds for a response from upstream, rather
than the default of 30 seconds.