Test bad responses from APIs
`encoding/json` has the property that it will parse things that aren’t
the expected shape and not highlight that the struct hasn’t been
populated with data.

This might not be what you want. There are various ways of handling
this. Verifying that the structure has a bare minimum is a simple way.

We should also verify the HTTP status codes of the responses coming back
as well.

Also ran `gofmt` on the code which reformatted and simplified some code.