Updgrade to go 1.6.2
This commit updates the .go-version file to 1.6.2 and updates the
dependencies.

`performanceplatform-client.go` has been renamed to
`performanceplatform-client-go` to allow it to be `go get`-ed and `gom
install`-ed

Regenerate `Gomfile` to include renamed performanceplatform-client and
get latest commits on dependencies.

`vendor` has been added to `.gitignore` as this replaces `_vendor` is now
the standard place for vendoring dependencies in go 1.6

Update `Makefile` to reflect the changes