Add `ApiRequest`
This provides higher level functionality to get a response from an API
and can be used across the API calls required within metadata-api. This
commit doesn't implement the `bearerToken` functionality required by
content API etc but the interface includes it as an argument to allow it
to be added in the future.