Add content_store package
This package queries the content store for a content item with the
supplied slug via the `GetArtefact` method. To maintain consistency
within the metadata API and to avoid breaking changes to the API response it
made sense to reuse the existing `Artefact`, `Detail` and `Part`
structs. This added a little bit of complexity as to match the content
store output with the existing structs required unmarshalling to a
`map[string]interface{}` and then a lot of type assertion.