Update format/doc-type logic after format retirement
The `format` field as a response from the content store has been
[retired][] and replaced with a combination of `schema_name` and
`document_type`.

This is not the cleanest solution: we're now exposing the document
type as `format` in the API response, which doesn't match up exactly
with the old values, and is now mis-named. However, the only app which
currently uses this API is info-frontend, and that doesn't use the
`format` value.

In addition, we should be changing info-frontend to talk to the
Performance Platform directly, rather than through metadata-api, which means we can retire metadata-api entirely.

[retired]: https://github.com/alphagov/govuk-rfcs/blob/fec5141c/rfc-041-separate-document-type-from-format.md